requirements.txt
# Airflow extras
apache-airflow==2.9.1
apache-airflow-providers-apache-spark==4.8.0
apache-airflow-providers-openlineage==1.8.0
apache-airflow-providers-slack==8.5.0

# Spark / Delta
pyspark==3.5.1
delta-spark==2.4.0

# Data Quality
great-expectations==0.18.19

# Monitoring
statsd==4.0.1

# Utils
python-dotenv==1.0.1
pydantic==2.7.1
loguru==0.7.2

# Testing
pytest==8.2.0
pytest-cov==5.0.0
pytest-mock==3.14.0
